chore(style): remove old style casts and cross sign comparisons
Enable warnings for both.

Fix #6010
Fix #6012

+++ src/model/chathistory.cpp
@@ -352,7 +352,7 @@ void ChatHistory::loadHistoryIntoSessionChatLog(ChatLogIdx start) const
     assert(getFirstIdx() == ChatLogIdx(0));
     auto messages = history->getMessagesForFriend(f.getPublicKey(), start.get(), end.get());
 
-    assert(messages.size() == end.get() - start.get());
+    assert(messages.size() == static_cast<int>(end.get() - start.get()));
     ChatLogIdx nextIdx = start;
 
     for (const auto& message : messages) {
